Rename WCCapturePaymentResponse to CapturePaymentResponse
